+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@
 - #114 Adds "How To Use Koffee" link in settings window
 - #111 Loads navigation history on startup
 - #115 Allows commenting out Recursive Search Exclusions
+- #124 Adds limit to length of BackStack and UndoStack
 - #70 Changes undo move/copy to undo individual file operations for more correct behavior when folders were merged
 - #123 Fixes path suggestions not being sorted by MRU
 - #126 Fixes search results to sort by current sort setting instead of always name ascending